Refactor how a DID for install is cached.
Previously, the DID was cached in the AJAX installation handler. This meant that only one DID could be cached, even during a bulk install (such as via WP-CLI).

Now, the DID is cached when its artifact's URL matches the package currently being passed for installation. Additionally, the cached DID is cleared after the package has been installed to prevent contamination.

Signed-off-by: costdev <79332690+costdev@users.noreply.github.com>