Bump the priority when renaming the downloaded package.
Repeated testing showed failures to rename the downloaded package correctly. This occurred at times for a package whether it used a release asset or not. While the exact cause hasn't been determined yet, bumping the priority resolved the issue.

Re-testing with the previous priority of 10, and possible investigation to find the cause, is welcome.

Signed-off-by: costdev <79332690+costdev@users.noreply.github.com>